Dependencies should be managed, not vendored
We had originally vendored a copy of ember-easyform, but it is now very,
very out of date. ember-cli has also started to disagree with the way
that it was being required, so the vendored copy would no longer load.

Use an ember-cli addon to include ember-easyform. This means we don’t
need to manually import EasyForm anymore.